test: add additional test cases for image analysis and error handling
